Add files via upload
Data files are for the Renko bars of different resolutions and brick sizes for SPY, Emini, VIX, and VVIX for  >20 years. All data are from extended hours available. 

Column headings describe the columns of data for each of the six indicators used to analyze the data and convert to graphic display in easy language from TradeStation and also in Multicharts. 
Hurst exponent is used to measuring the hurst component as described by MARKO TAVORA